requirements.txt
numpy
matplotlib
tqdm
torch
torchvision
ipywidgets
tensorflow
Pillow
keras
shap
scikit-image